news/requirements.txt
fastapi
uvicorn
groq
pinecone-client
requests
python-dotenv
jinja2
